nix-daemon/wire: add string serialization test from cppnix
This commit adapts the first test case and known-good test vector
from cppnix's tests/unit/libstore/worker-protocol.cc, as of commit
91b6833686a6a6d9eac7f3f66393ec89ef1d3b57.

Unfortunately the cppnix protocol tests are written directly in C++,
which means they can't really be run directly in Rust.  This is in
contrast to the cppnix libexpr tests, which are just a large pile of
.nix files and their corresponding expected outputs when evaluated.

The cppnix codebase is licensed LGPL, which is allegededly
compatible with the EUPL.  To clarify the signoff below, I was the
one who adapted these tests.

Signed-off-by: Adam Joseph <adam@westernsemico.com>